Major update. Refactor solution, add HNSW, refactor buildprompt
