+++ Search/Pages/Index.razor
@@ -1,4 +1,5 @@
-﻿@page "/"
+﻿@using SharedLib.Models
+@page "/"
 
 <PageTitle>Azure Cosmos DB & Azure OpenAI</PageTitle>
 <main class="h-100 d-flex flex-row justify-content-start">
