+++ Search/Pages/NavMenu.razor
@@ -1,5 +1,6 @@
 ﻿@using Search.Constants
 @using Search.Services
+@using SharedLib.Models
 @inject ChatService chatService
