+++ Search/_Imports.razor
@@ -6,7 +6,7 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
+@using SharedLib
 @using Search
-@using Search.Models
 @using Search.Components
 @using Search.Shared
