+++ Search/appsettings.json
@@ -10,15 +10,17 @@
   "OpenAi": {
     "Endpoint": "",
     "Key": "",
-    "EmbeddingsDeployment": "embeddings",
-    "CompletionsDeployment": "completions",
-    "MaxConversationTokens": "100",
-    "MaxCompletionTokens": "500"
+    "EmbeddingsDeployment": "",
+    "CompletionsDeployment": "",
+    "MaxConversationTokens": "1000",
+    "MaxCompletionTokens": "2000",
+    "MaxEmbeddingTokens": "8000"
   },
   "MongoDb": {
     "Connection": "",
     "DatabaseName": "retaildb",
-    "CollectionNames": "product,customer,vectors,completions",
-    "MaxVectorSearchResults": "10"
+    "CollectionNames": "products, customers, salesOrders, completions",
+    "MaxVectorSearchResults": "10",
+    "VectorIndexType": "hnsw" // ivf, hnsw
   }
 }
