+++ SharedLib/Models/Session.cs
@@ -2,7 +2,7 @@ using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 using Newtonsoft.Json;
 
-namespace Search.Models;
+namespace SharedLib.Models;
 
 public record Session
 {
