+++ SharedLib/SharedLib.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="MongoDB.Driver" Version="2.*" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.*" />
+		<PackageReference Include="Azure.AI.OpenAI" Version="1.0.0-beta.*" />
+		<PackageReference Include="SharpToken" Version="1.2.12" />
+	</ItemGroup>
+
+</Project>
