+++ VectorSearchAiAssistantMongoDBvCore.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharedLib", "..\AzureAiStudio-MongoDBvCore\SharedLib\SharedLib.csproj", "{FF807B70-8F25-4A28-A017-3DA79ADD4DB2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,6 +31,10 @@ Global
 		{E8C2FE79-83B4-460F-BACD-467E9200C9DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8C2FE79-83B4-460F-BACD-467E9200C9DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8C2FE79-83B4-460F-BACD-467E9200C9DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FF807B70-8F25-4A28-A017-3DA79ADD4DB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF807B70-8F25-4A28-A017-3DA79ADD4DB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF807B70-8F25-4A28-A017-3DA79ADD4DB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF807B70-8F25-4A28-A017-3DA79ADD4DB2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
