+++ Vectorize/Vectorize.csproj
@@ -26,6 +26,9 @@
 	  <PackageReference Include="MongoDB.Driver" Version="2.*" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SharedLib\SharedLib.csproj" />
+  </ItemGroup>
+  <ItemGroup>
     <None Update="host.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
